User class
Finished most of the user class as a demo--I wouldn't take any problems
to heart, I didn't spend that much time on it since I figure we're going
to talk about it. The demo code is in the MainWindow class--it creates a
user, displays the fields, updates a field, then displays the updated
field. It will throw an exception if it's been run before since the
username needs to be unique (just delete the .db if you want to run it
again, it will recreate it).