Refactor listviews
Need to distinguish between a list view and a search view--not all list
views will need the search box field (e.g. selecting an address for a
transaction). The search box is too entwined with the ListView code, so
it's probably less trouble to start a new type of ListView from scratch.